+++ demisto_sdk/commands/validate/tests/tools_test.py
@@ -2,7 +2,6 @@ import pytest
 
 from demisto_sdk.commands.validate.tests.test_tools import create_playbook_object
 from demisto_sdk.commands.validate.tools import (
-    check_timestamp_format,
     collect_all_inputs_from_inputs_section,
     collect_all_inputs_in_use,
 )
@@ -88,26 +87,3 @@ def test_collect_all_inputs_from_inputs_section(content_item, expected_result):
         - A set with all inputs defined in the inputs section with no duplicates or empty spaces are returned.
     """
     assert collect_all_inputs_from_inputs_section(content_item) == expected_result
-
-
-def test_check_timestamp_format():
-    """
-    Given
-    - timestamps in various formats.
-
-    When
-    - Running check_timestamp_format on them.
-
-    Then
-    - Ensure True for iso format and False for any other format.
-    """
-    good_format_timestamp = "2020-04-14T00:00:00Z"
-    missing_z = "2020-04-14T00:00:00"
-    missing_t = "2020-04-14 00:00:00Z"
-    only_date = "2020-04-14"
-    with_hyphen = "2020-04-14T00-00-00Z"
-    assert check_timestamp_format(good_format_timestamp)
-    assert not check_timestamp_format(missing_t)
-    assert not check_timestamp_format(missing_z)
-    assert not check_timestamp_format(only_date)
-    assert not check_timestamp_format(with_hyphen)
