+++ demisto_sdk/commands/validate/validators/PA_validators/PA107_missing_field_in_pack_metadata.py
@@ -31,7 +31,9 @@ class MissingFieldInPackMetadataValidator(BaseValidator[ContentTypes]):
             for content_item in content_items
             if (
                 missing_fields := [
-                    field for field in PACK_METADATA_FIELDS if field not in content_item.pack_metadata_dict  # type: ignore
+                    field
+                    for field in PACK_METADATA_FIELDS
+                    if field not in content_item.pack_metadata_dict  # type: ignore[operator]
                 ]
             )
         ]
@@ -41,7 +43,7 @@ class MissingFieldInPackMetadataValidator(BaseValidator[ContentTypes]):
         content_item: ContentTypes,
     ) -> FixResult:
         missing_fields = [
-            field for field in PACK_METADATA_FIELDS if field not in content_item.pack_metadata_dict  # type: ignore
+            field for field in PACK_METADATA_FIELDS if field not in content_item.pack_metadata_dict  # type: ignore[operator]
         ]
         return FixResult(
             validator=self,
