+++ demisto_sdk/commands/validate/validators/PA_validators/PA111_empty_metadata_fields.py
@@ -29,7 +29,7 @@ class EmptyMetadataFieldsValidator(BaseValidator[ContentTypes]):
             for content_item in content_items
             if (
                 missing_fields := [
-                    field for field in PACK_METADATA_MANDATORY_FILLED_FIELDS if field in content_item.pack_metadata_dict and (field_val := content_item.pack_metadata_dict[field]) and "" in field_val  # type: ignore
+                    field for field in PACK_METADATA_MANDATORY_FILLED_FIELDS if field in content_item.pack_metadata_dict and (field_val := content_item.pack_metadata_dict[field]) and "" in field_val  # type: ignore[operator, index]
                 ]
             )
         ]
